MedAgentGym/medAgentGym/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-aiplatform
python-dotenv
google-cloud-storage
pandas
